requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
discord.py==2.2.2
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
